chore: rewrite reporter-helpers to typescript
